feat: work on vertical notifications
