+++ .github/workflows/actions/send-rocketchat-notification/action.yml
@@ -28,6 +28,6 @@ runs:
         mention: "jason.leach @cvarjao @wade.king @bryce.mcmath"
         mention_if: always
         channel: "#bc-digital-wallet-core-team"
-        url: ${{ inputs.webhook_url }}
         commit: true
+        url: ${{ inputs.webhook_url }}
         token: ${{ inputs.webhook_token }}
