+++ app/src/hooks/notifications.ts
@@ -1,4 +1,3 @@
-import { AnonCredsCredentialMetadataKey } from '@aries-framework/anoncreds/build/utils/metadata'
 import {
   CredentialExchangeRecord as CredentialRecord,
   CredentialState,
@@ -6,12 +5,8 @@ import {
   ProofState,
 } from '@aries-framework/core'
 import { useCredentialByState, useProofByState } from '@aries-framework/react-hooks'
-import { useStore } from 'aries-bifold'
 import { CredentialMetadata, customMetadata } from 'aries-bifold/App/types/metadata'
 import { ProofCustomMetadata, ProofMetadata } from 'aries-bifold/verifier'
-
-import { getInvitationCredentialDate, showBCIDSelector } from '../helpers/BCIDHelper'
-import { BCState } from '../store'
 
 interface CustomNotification {
   type: 'CustomNotification'
@@ -25,7 +20,6 @@ interface Notifications {
 }
 
 export const useNotifications = (): Notifications => {
-  const [store] = useStore<BCState>()
   const offers = useCredentialByState(CredentialState.OfferReceived)
   const proofsRequested = useProofByState(ProofState.RequestReceived)
   const proofsDone = useProofByState([ProofState.Done, ProofState.PresentationReceived]).filter(
@@ -44,22 +38,7 @@ export const useNotifications = (): Notifications => {
     }
   })
 
-  const credentials = [
-    ...useCredentialByState(CredentialState.CredentialReceived),
-    ...useCredentialByState(CredentialState.Done),
-  ]
-  const credentialDefinitionIDs = credentials.map(
-    (c) => c.metadata.data[AnonCredsCredentialMetadataKey].credentialDefinitionId as string
-  )
-  const invitationDate = getInvitationCredentialDate(credentials, true)
-  const custom: CustomNotification[] =
-    showBCIDSelector(credentialDefinitionIDs, true) &&
-    invitationDate &&
-    !store.dismissPersonCredentialOffer.personCredentialOfferDismissed
-      ? [{ type: 'CustomNotification', createdAt: invitationDate, id: 'custom' }]
-      : []
-
-  const notifications = [...offers, ...proofsRequested, ...proofsDone, ...revoked, ...custom].sort(
+  const notifications = [...offers, ...proofsRequested, ...proofsDone, ...revoked].sort(
     (a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()
   )
