+++ app/src/index.ts
@@ -13,7 +13,8 @@ import bundles from './assets/branding/credential-branding'
 import AddCredentialButton from './components/AddCredentialButton'
 import AddCredentialSlider from './components/AddCredentialSlider'
 import EmptyList from './components/EmptyList'
-import HomeContentView from './components/HomeContentView'
+import HomeFooterView from './components/HomeFooterView'
+import HomeHeaderView from './components/HomeHeaderView'
 import { useNotifications } from './hooks/notifications'
 import en from './localization/en'
 import fr from './localization/fr'
@@ -38,7 +39,8 @@ const configuration: ConfigurationContext = {
   pages,
   splash: Splash,
   terms: Terms,
-  homeContentView: HomeContentView,
+  homeHeaderView: HomeHeaderView,
+  homeFooterView: HomeFooterView,
   credentialListHeaderRight: AddCredentialButton,
   credentialListOptions: AddCredentialSlider,
   credentialEmptyList: EmptyList,
